Development Update 2
General:
Added ability to import a large greyscale tomography morphology image,
analyze and assign sites site types, and extract image segments to form
a morphology set that is then analyzed to determine the morphological
characteristics of the set.  Code is designed so that sites types 1 and
2 are the primary donor and acceptor phases and any additional types are
secondary.  An implication of this design is that tortuosity is only
calculated for types 1 and 2.

main:
-added parameters for performing the tomogram import and analysis
-updated importParameters function to read new parameters from the
parameter file
-added acceptance of a new command line argument '-import' that
specifies a regular morphology file import
-added acceptance of a new command line argument '-importTomogram' that
tells the program to expect a binary format tomogram file
-added code block to create a Morphology object with a tomogram data set
and then extract a vector of Morphology objects that consist of
subsegments of the total tomogram and multiple random variants can also
be created of each subsegment
-moved array_median function to Utils and coverted it to a more general
template function
-moved much of the details of the domain size and anisotropy calculation
procedure into the Morphology class instead of complicating main
-moved calculations of the interfacial area to volume ratio, interfacial
volume fraction, and island volume fraction into the Morphology class
-corrected bug with analysis_file output to not print out the domian
size and tortuosity information when the respective calculations are not
enabled

parameters_default:
-added parameter section and parameters for the tomogram import feature

Morphology class:
-added definition of new data struct TomogramImportParams to be used to
store collect and pass all tomogram import and analysis parameters
-added several new constructors, one creating a blank Morphology object,
and one creating a Morphology object using an existing Lattice object
-updated class to handle an arbitratry number of different site types,
allowing morphologies to go beyond binary blends such as ternary blend
and even more complex blends
-updated calculateAnisotropies to perform tasks that were previously
handled in main, primarily incrementing the cutoff radius for the
correlation length calculation until enough data is obtained to
determine the anisotropy
-added calculateCorrelationDistances function to handle tasks previously
handled in main, primarily incrementing the cutoff radius for the
correlation length calculation until enough data is obtained to
determine the domain size
-updated calculateCorrelationDistance function to determine the
correlation length from when the correlation function reaches 1/e value,
and the domain size is now determined as 2 times the correlation length.
This allow domains size calculation for less periodic Morphologies that
have correlation functions that do not cross the mix fraction level
-changed and updated calculateInterfacialArea function to
calculateInterfacialAreaVolumeRatio function so that volume
normalization does not need to be performed in main
-changed and updated calculateInterfacialVolume function to
calculateInterfacialVolumeFraction function so that volume normalization
does not need to be performed in main
-changed and updated getIslandVolume function to getIslandVolumeFraction
function so that volume normalization does not need to be performed in
main
-updated createRandomMorphology function to accept a vector of
mix_fractions to handle the case where there are more than 2 types of
sites in the blend
-updated getMixFraction to accept a site type input to output the volume
fraction of the input site type relative to the total volume to handle
the case where there are more than 2 tpyes of sites in the blend
-added importTomogramMorphologyFile function to handle binary file
reading and analysis of tomogram data using the TomogramImportParams
parameter struct and return a vector of Morphology objects
-updated all characteristcs that were specific to a site type to now be
a vector of characteristics with the vector size equal to the number of
different site types
-added addSiteType function to be used to when a new site type is added
to the Morphology, which updates the size of all morphology
characteristic vectors
-updated getSiteSampling function to accept a site type input and return
only randomly selecting sites that have a matching site type
-updated outputMorphologyFile so that morphology data files contain
boundary condition details, number of site types, and domain size and
mix fractions for each site type
-updated importMorphology to detect v4 morphology and read in the
additional information contained in the new version data format

Lattice class:
-added extractSublattice function to extract a small lattice with the
defined coordinates and size from the larger lattice

Utils:
-added calculateCumulativeHist function to integrate an x-y data set and
return the cumulative histogram
-added template function array_median that determines the median value
of a data set